requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.2
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
stripe==5.4.0